This commit is to note that I completed the API calls and passing information between modules.
I also added the new items into state where
necessary and passed props down into the relevant
modules.

Lastly I found an error in the seeding for the
Review module in which it was not seeding all of
the restaurants.  I fixed this and will pass the
information on to Alvin in case he has not fixed
it in the current version of his repo.

Next task to upload to AWS.